+++ src/features/notifications/index.js
@@ -1,16 +1,16 @@
 import React from 'react'
 import { Row, Col } from 'antd'
-import { NotificationsContainer } from './styled-components'
-import { NotificationItem } from './NotificationItem'
-import listData from '__mocks__/mock-data'
+import NotificationsContainer from './styled-components'
+import NotificationItem from './NotificationItem'
+import { notificationsData } from '__mocks__/mock-data'
 
 const Notifications = () => (
   <Row type="flex" justify="center" align="middle">
-    <Col span={24}>
+    <Col span={20}>
       <NotificationsContainer
         itemLayout="vertical"
         size="large"
-        dataSource={listData}
+        dataSource={notificationsData}
         renderItem={item => <NotificationItem item={item} />}
       />
     </Col>
